feat: add kafka config admin

+++ src/main/kotlin/com/example/eventdriven/application/infrastructure/config/data/KafkaConfigData.kt
@@ -12,5 +12,5 @@ data class KafkaConfigData(
     val topicName: String,
     val topicNameToCreate: List<String>,
     val numOfPartitions: Int,
-    val replicationFactor: Int,
+    val replicationFactor: Short,
 )
